moves LibvirtConnection into libs/libvirt
